MCR-3657 rename module to mycore-solr-api
