+++ mycore-bom/pom.xml
@@ -206,7 +206,7 @@
       </dependency>
       <dependency>
         <groupId>org.mycore</groupId>
-        <artifactId>mycore-solr-restapi</artifactId>
+        <artifactId>mycore-solr-api</artifactId>
         <version>${project.version}</version>
       </dependency>
       <dependency>
