+++ mycore-meta/pom.xml
@@ -168,7 +168,7 @@
     </dependency>
     <dependency>
       <groupId>org.mycore</groupId>
-      <artifactId>mycore-solr-restapi</artifactId>
+      <artifactId>mycore-solr-api</artifactId>
     </dependency>
     <dependency>
       <groupId>org.mycore</groupId>
