+++ mycore-solr-api/LICENSE.txt
