+++ mycore-solr-api/pom.xml
@@ -27,8 +27,8 @@
     <version>2026.06.0-SNAPSHOT</version>
     <relativePath>../pom.xml</relativePath>
   </parent>
-  <artifactId>mycore-solr-restapi</artifactId>
-  <name>MyCoRe Solr REST API</name>
+  <artifactId>mycore-solr-api</artifactId>
+  <name>MyCoRe Solr API</name>
   <dependencies>
     <dependency>
       <groupId>jakarta.ws.rs</groupId>
