+++ mycore-solr-api/src/main/resources/components/solr-api/config/mycore.properties
@@ -1,2 +1,2 @@
-MCR.Solr.Rest.Resource.Packages=org.mycore.solr.rest
+MCR.Solr.Api.Resource.Packages=org.mycore.solr.api
 MCR.Jersey.Resource.ApplicationPaths=%MCR.Jersey.Resource.ApplicationPaths%,api/solr
